edited shedule
